Adds `stream!/3` interface
This is a rough draft of what I think the stream implementation should
be like. I do not know how I am going to test the error case where the
file can not be found with any of the adapters. The exception happens in
a supervised storage implementation that is inside of a macro for the
adapter cases.